% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cdc_authored.R
\name{check_for_packagesrds}
\alias{check_for_packagesrds}
\title{Check for the CRAN packages records}
\usage{
check_for_packagesrds(
  url = "http://cran.r-project.org/web/packages/packages.rds"
)
}
\arguments{
\item{url}{Required. Location to use when searching for \code{packages.rds}.
Defaults to the expected location of
\url{http://cran.r-project.org/web/packages/packages.rds}.}
}
\value{
Returns the url if the file is found, otherwise returns an error.
}
\description{
Checks if the \code{packages.rds} file is available in CRAN.
}
\examples{
check_for_packagesrds()

}
