% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cdc_authored.R
\name{get_author_info}
\alias{get_author_info}
\title{Get author information}
\usage{
get_author_info(.data, elements, regex = "Disease Control|CDC|cdc[.]gov")
}
\arguments{
\item{.data}{Data frame with package and author information.}

\item{elements}{Which elements to extract.}

\item{regex}{Which authors to match.}
}
\value{
The person fields requested for the authors matching the regex.
}
\description{
Extracts person fields from the authors.
}
